requirements.txt
langchain==0.0.316
openai==0.28.1
streamlit
tiktoken
python-dotenv
unstructured
pinecone-client
pypdf
sentence_transformers